Cookstyle 5.10
Fairly minor updates for issues found in Cookstyle 5.10. I also modernized the .rubocop.yml file a bit for cookstyle and newer rubocop releases where some cops have been removed and other renamed.

Note: I updated all the 'Apache' license values to be 'Apache-2.0'. I assume that's what the intention was given the license of the repo. I noticed that most of your file headers are all rights reserved, which is going to cause issues for a lot of people trying to use these cookbooks that require approval from their legal teams. You might want to drop the Apache headers on the recipes in place of the all rights message.

Signed-off-by: Tim Smith <tsmith@chef.io>